requirements/requirements.txt
Django==1.8.8
https://github.com/hendrix/hendrix/archive/master.zip
django-encrypted-fields==1.1.1
django-phonenumber-field
factory_boy
mock
mandrill
psycopg2
newrelic
raven
bleach

pep8
